# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Replay já 2.0 — API
  version: 2.0.0-draft.1
  summary: Contrato entre o relay de gravação, o botão da quadra, o app do atleta e o painel do parceiro.
  description: |
    Contrato da API do Replay já 2.0.

    **Arquitetura**: a câmera IP da quadra **empurra RTMP** direto para um *relay* nosso
    (Lightsail, `sa-east-1`), que grava 24/7 em fMP4 com `ffmpeg -c copy` e corta o clipe a
    partir do índice. Não há computador de borda na arena. O botão é apenas um evento HTTP.

    | Superfície | Host | Autenticação |
    |---|---|---|
    | **Gatilho — botão físico** | `replayja.com.br/api/v1` | token no caminho (`/triggers/b/{buttonToken}`) |
    | **Gatilho — botão virtual** | `replayja.com.br/api/v1` | `userSession` |
    | **App do atleta** | `replayja.com.br/api/v1` | `userSession` (cookie HMAC próprio) |
    | **Painel do parceiro** | `replayja.com.br/api/v1` | `userSession` + `partner_admin` ativo |
    | **Relay → nuvem** | `replayja.com.br/api/relay` | `relayKey` (header `x-relay-key`) |
    | **Nuvem → relay** | `relay-1.replayja.com.br` | `relayToken` (HMAC no caminho) |

    **Quem manda em quem** (padrão herdado do relay v2 do Sentinela, em produção desde
    29/08/2026): o app **não escreve** no relay. O relay pergunta — lista de câmeras, fila de
    jobs de clipe — e a resposta é sempre uma *lista*, nunca um comando. A única chamada
    nuvem → relay é o `POST /jobs`, que apenas **acorda** o relay para buscar a fila mais cedo;
...

...
servers:
...
security:
  - userSession: []

# ---------------------------------------------------------------------------
paths:
...
  /triggers:
    post:
      tags: [Gatilhos, Clipes]
      summary: Gatilho do botão virtual (usuário logado no site)
      operationId: createTrigger
      description: |
        Mesmo caminho do botão físico, com sessão de usuário em vez de token no caminho.
        Como o relay grava 24/7, o botão virtual **não depende de nenhuma conexão persistente
        com a arena** — é uma linha no banco e uma chamada ao relay. O long-polling e a fila
        de comandos que existiam no desenho com computador de borda deixaram de existir.

        Modos de falha:

        | Situação | HTTP | Mensagem ao usuário |
        |---|---|---|
        | Câmera com cobertura normal | `202` | "Lance salvo! Aparece aqui em instantes." |
        | Cooldown (< 8 s do último gatilho na quadra) | `429` | "Aguarde alguns segundos." |
        | Câmera sem segmento há > 60 s | `409` | "A câmera desta quadra está fora do ar." |
        | Relay indisponível | `503` | "Estamos com um problema técnico, tente em instantes." |

        Rate limit: 10 gatilhos/hora por usuário por quadra.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [courtId]
              properties:
                courtId: { type: string, format: uuid }
                preRollSeconds: { type: integer, minimum: 5, maximum: 60, default: 24 }
                postRollSeconds: { type: integer, minimum: 0, maximum: 30, default: 1 }
      responses:
        '202':
          description: Job de clipe criado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TriggerAck' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409':
          description: Câmera sem cobertura no momento
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503':
          description: Relay indisponível
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
...
components:
  securitySchemes:
    userSession:
      type: apiKey
      in: cookie
      name: session
      description: |
        Cookie de sessão assinado com HMAC-SHA256 (`<payload base64url>.<assinatura>`),
        `HttpOnly`, `Secure`, `SameSite=Lax`, validade de 400 dias com renovação deslizante
        a cada 7 dias de uso. Padrão portado do Sentinela (`lib/session.ts`).

        O payload é `{uid, email, exp}` e **não carrega autorização**: papel de admin de arena
        é consultado no banco na requisição de painel, para que remover um admin tenha efeito
        imediato em vez de esperar o cookie expirar.

        Não há tabela de sessão, nem provedor de identidade externo. Login por OTP de 6 dígitos
        (desafio também em cookie HMAC, de 10 min) ou por Google via OIDC.
    relayKey:
      type: apiKey
      in: header
      name: x-relay-key
      description: |
        Segredo por **relay** (não por câmera, não por arena), de 32 bytes, guardado em
        `/etc/replayja/rec.env` (modo 600, root) e com o SHA-256 no Postgres.

        Dá acesso a: a lista de câmeras daquele relay (incluindo as chaves RTMP), a fila de
        jobs daquele relay e as URLs de upload. **Não** dá acesso a nenhum dado de usuário.
        Rotação: mudar no painel, aplicar no relay, a chave antiga vale por 10 min.
    relayToken:
      type: apiKey
      in: header
      name: x-relay-key
      description: |
        Mesma credencial, na direção oposta: a API usa este cabeçalho ao chamar o relay
        (`POST /jobs`, `GET /stats`). Nas rotas `/t/{relayToken}/...` o token vai no **caminho**,
        porque ele é colocado num `<img src>` ou num link de download e não há como mandar
        cabeçalho — é um HMAC de escopo estreito (uma câmera, um intervalo) e validade curta,
        validado localmente pelo `auth-sidecar` do relay. Validar na nuvem viraria centenas de
        invocações por minuto.
    buttonToken:
      type: apiKey
      in: query
...
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        UUID gerado pelo cliente. Requisições repetidas com a mesma chave, no mesmo endpoint e
        dentro de 24h, devolvem a resposta original sem repetir o efeito. Ver
        `docs/api/README.md` §5.
      schema: { type: string, format: uuid }
  responses:
    Unauthorized:
      description: Credencial ausente, inválida ou expirada
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    TooManyRequests:
      description: Limite de taxa excedido
      headers:
        Retry-After:
          schema: { type: integer }
        RateLimit:
          description: RFC 9331 (`limit=…, remaining=…, reset=…`)
          schema: { type: string }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      description: RFC 9457 — Problem Details for HTTP APIs.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          examples: ['https://replayja.com.br/problems/clip-expired']
        title: { type: string, examples: ['Clipe expirado'] }
        status: { type: integer }
        detail:
          type: string
          description: Mensagem em pt-BR, pronta para exibir ao usuário.
        instance: { type: string }
        traceId: { type: string, description: 'Correlaciona com a linha em `app_error` e com o log da Vercel.' }
        errors:
          type: array
          description: Erros de validação campo a campo.
          items:
            type: object
            properties:
              field: { type: string }
              code: { type: string }
              message: { type: string }
...
    TriggerAck:
      type: object
      required: [accepted]
      properties:
        accepted: { type: boolean }
        triggerEventId: { type: string, format: uuid, nullable: true }
        clipId: { type: string, format: uuid, nullable: true }
        outcome:
          type: string
          enum: [accepted, rejected_cooldown, rejected_no_coverage, rejected_camera_unknown, rejected_button_revoked, rejected_relay_down]
        arrivalAt: { type: string, format: date-time }
        pressEstimatedAt:
          type: string
          format: date-time
          description: '`arrivalAt − button.wakeLatencyMs`. É o instante usado para montar a janela.'
        appliedLatencies:
          type: object
          properties:
            buttonWakeLatencyMs: { type: integer }
            cameraOriginLagMs: { type: integer }
        window:
          type: object
          properties:
            deliverFrom: { type: string, format: date-time }
            deliverTo: { type: string, format: date-time }
...